Update existing ts tests to V2
